requirements.txt
blend-modes==2.1.0
colorcorrect==0.9
numpy==1.18.1
Pillow==7.0.0
six==1.14.0
